refactor message
